+++ pkg/dashboard/resources/pvcs/service.go
@@ -1,16 +1,18 @@
-// Copyright 2025 Juicedata Inc
-//
-// Licensed under the Apache License, Version 2.0 (the "License");
-// you may not use this file except in compliance with the License.
-// You may obtain a copy of the License at
-//
-//     http://www.apache.org/licenses/LICENSE-2.0
-//
-// Unless required by applicable law or agreed to in writing, software
-// distributed under the License is distributed on an "AS IS" BASIS,
-// WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
-// See the License for the specific language governing permissions and
-// limitations under the License.
+/*
+ Copyright 2025 Juicedata Inc
+
+ Licensed under the Apache License, Version 2.0 (the "License");
+ you may not use this file except in compliance with the License.
+ You may obtain a copy of the License at
+
+     http://www.apache.org/licenses/LICENSE-2.0
+
+ Unless required by applicable law or agreed to in writing, software
+ distributed under the License is distributed on an "AS IS" BASIS,
+ WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ See the License for the specific language governing permissions and
+ limitations under the License.
+*/
 
 package pvcs
 
@@ -24,9 +26,9 @@ import (
 )
 
 type ListPVCResult struct {
-	Total         int                            `json:"total"`
-	ContinueToken string                         `json:"continueToken"`
-	PVCs          []corev1.PersistentVolumeClaim `json:"pvcs"`
+	Total    int                            `json:"total,omitempty"`
+	Continue string                         `json:"continue,omitempty"`
+	PVCs     []corev1.PersistentVolumeClaim `json:"pvcs"`
 }
 
 type PVCService interface {
