+++ pkg/dashboard/resources/pvs/pv_service.go
@@ -1,16 +1,18 @@
-// Copyright 2025 Juicedata Inc
-//
-// Licensed under the Apache License, Version 2.0 (the "License");
-// you may not use this file except in compliance with the License.
-// You may obtain a copy of the License at
-//
-//     http://www.apache.org/licenses/LICENSE-2.0
-//
-// Unless required by applicable law or agreed to in writing, software
-// distributed under the License is distributed on an "AS IS" BASIS,
-// WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
-// See the License for the specific language governing permissions and
-// limitations under the License.
+/*
+ Copyright 2025 Juicedata Inc
+
+ Licensed under the Apache License, Version 2.0 (the "License");
+ you may not use this file except in compliance with the License.
+ You may obtain a copy of the License at
+
+     http://www.apache.org/licenses/LICENSE-2.0
+
+ Unless required by applicable law or agreed to in writing, software
+ distributed under the License is distributed on an "AS IS" BASIS,
+ WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ See the License for the specific language governing permissions and
+ limitations under the License.
+*/
 
 package pvs
 
@@ -67,8 +69,8 @@ func (s *pvService) ListPVs(c *gin.Context) (*ListPVPodResult, error) {
 		return nil, err
 	}
 	result := &ListPVPodResult{
-		PVs:           pvs,
-		ContinueToken: nextContinue,
+		PVs:      pvs,
+		Continue: nextContinue,
 	}
 	return result, nil
 }
